Move WarningTitle out of components/common
